jlkaus: debug dftgen

+++ pipe-stuff/dftgen.C
@@ -158,7 +158,7 @@ int main(int argc, char* argv[]) {
 		      strcat(sysspace,".dft");	      
 
 		      printf("O %1d, BF %.3f, NC %06.2f, M %1d, MC %06.2f, ", octave, base_freq, note_cents, mutation, mutation_cents);
-		      printf("D %d, DC %06.2f, FC %06.2f, F %.3f, K %.0f, ", detune_c, detune_cents, f_cents, fundamental, k_max);
+		      printf("D %d, DC %06.2f, FC %06.2f, F %.3f, K %d, ", detune_c, detune_cents, f_cents, fundamental, (int)k_max);
 		      printf("%s\n", spectrum_fn);
 
 		      FILE* sf = fopen(sysspace, "wb");
